use temporary files to be smarter than linters
Both go vet and golint consider a file a go source if the extension is ".go" or
if the file doesn't have parse error (whatever the name).
This patch uses a temporary file with a ".go" extension when needed so that the
linter also report parse errors in files without extension.